requirements.txt
pyttsx3~=2.90
requests~=2.27.1
playsound~=1.2.2
wikipedia~=1.4.0
gTTS~=2.2.3
Pillow~=9.1.0
selenium~=4.1.2